chore: moves `repositories` and `services` folders from `/app` to `/`
